pyproject.toml
[build-system]
requires = ["hatchling>=1.24.2"]
build-backend = "hatchling.build"

[project]
name = "emr-service"
version = "0.1.0"
description = "Electronic medical record backend built with FastAPI"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14.0",
  "asyncpg>=0.30.0",
  "email-validator>=2.2.0",
  "fastapi>=0.115.0",
  "greenlet>=3.1.1",
  "pydantic-settings>=2.6.1",
  "prometheus-client>=0.21.1",
  "pyjwt>=2.10.1",
  "sqlalchemy[asyncio]>=2.0.36",
  "uvicorn[standard]>=0.32.1",
]

[project.optional-dependencies]
dev = [
  "aiosqlite>=0.20.0",
  "httpx>=0.28.1",
  "pytest>=8.3.4",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=6.0.0",
  "ruff>=0.8.4",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]
